third_party/WebKit/LayoutTests/fast/forms/date/input-date-validation-message-expected.txt

Test for validationMessage IDL attribute for <input type=date>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


No message
PASS testIt("", "", "") is ""
Value missing
PASS testIt("", "", "") is "Please fill out this field."
Type mismatch
PASS testIt("foo", "", "") is ""
Range overflow
PASS testIt("1982-11-02", "", "1970-12-31") is "Value must be less than or equal to 12/31/1970."
Range underflow
PASS testIt("1982-11-02", "1990-05-25", "1990-12-24") is "Value must be greater than or equal to 05/25/1990."
Step mismatch
PASS testIt("1982-11-02", "", "", "123") is "Please enter a valid value. The two nearest valid values are 07/07/1982 and 11/07/1982."
PASS successfullyParsed is true

TEST COMPLETE
